Removed all modals and replaced them with generic notification modal
