+++ source/Pages/Notification/Views/Transfer/hooks/useRequests.jsx
@@ -9,6 +9,9 @@ import PlugController from '@psychedelic/plug-controller';
 import { Principal } from '@dfinity/agent';
 import { validatePrincipalId } from '@shared/utils/ids';
 import { DEFAULT_FEE } from '@shared/constants/addresses';
+import { setAssets, setTransactions } from '@redux/wallet';
+import { HANDLER_TYPES, sendMessage } from '@background/Keyring';
+import { useDispatch } from 'react-redux';
 
 const portRPC = new PortRPC({
   name: 'transfer-port',
@@ -18,10 +21,12 @@ const portRPC = new PortRPC({
 
 portRPC.start();
 
-const useRequests = (incomingRequests, callId, portId) => {
+const useRequests = (incomingRequests, callId, portId, icpPrice) => {
   const { t } = useTranslation();
   const [currentRequest, setCurrentRequest] = useState(0);
   const [requests, setRequests] = useState(incomingRequests);
+  const [loading, setLoading] = useState(false);
+  const dispatch = useDispatch();
 
   const [response, setResponse] = useState([]);
 
@@ -49,8 +54,25 @@ const useRequests = (incomingRequests, callId, portId) => {
 
   useEffect(async () => {
     if (requests.length === 0) {
-      await portRPC.call('handleRequestTransfer', [response, callId, portId]);
-      window.close();
+      setLoading(true);
+      if (response?.[0]?.status === 'declined') {
+        await portRPC.call('handleRequestTransfer', [{ ok: false, error: 'The transaction was rejected' }, callId, portId]);
+        window.close();
+      } else {
+        sendMessage({
+          type: HANDLER_TYPES.SEND_ICP,
+          params: response?.[0],
+        }, async (sendResponse) => {
+          const { error, assets: keyringAssets, transactions } = sendResponse || {};
+          if (!error) {
+            dispatch(setAssets(keyringAssets));
+            dispatch(setTransactions({ ...transactions, icpPrice }));
+          }
+          await portRPC.call('handleRequestTransfer', [{ ok: !error, error, response: sendResponse }, callId, portId]);
+          setLoading(false);
+          window.close();
+        });
+      }
     }
   }, [requests]);
 
@@ -94,11 +116,11 @@ const useRequests = (incomingRequests, callId, portId) => {
   data.push(...[
     {
       label: t('common.fee'),
-      component: <DataDisplay value={<AssetFormat value={requests[currentRequest].args?.fee || DEFAULT_FEE} asset={CURRENCIES.get('ICP').value} />} />,
+      component: <DataDisplay value={<AssetFormat value={requests[currentRequest]?.args?.fee || DEFAULT_FEE} asset={CURRENCIES.get('ICP').value} />} />,
     },
     {
       label: t('common.memo'),
-      component: <DataDisplay value={requests[currentRequest].args?.memo || t('common.null')} />,
+      component: <DataDisplay value={requests[currentRequest]?.args?.memo || t('common.null')} />,
     },
   ]);
 
@@ -111,6 +133,7 @@ const useRequests = (incomingRequests, callId, portId) => {
     handleRequest,
     handleDeclineAll,
     principalId,
+    loading,
   };
 };
