--- source/Pages/Transfer/constants/index.js
@@ -1,9 +1,0 @@
-const SIZES = {
-  width: 436,
-  dataHeightBig: 632,
-  dataHeightSmall: 549,
-  detailsHeightBig: 649,
-  detailHeightSmall: 564,
-};
-
-export default SIZES;
